\name{SpecLibDolphin2014-package}
\alias{SpecLibDolphin2014-package}
\docType{package}
\title{
Mass Spectral Library Pacific Bottlenose Dolphin Blubber 2014
}
\description{
Contains the code and mass spectral library described in: Nontargeted analysis reveals hundreds of unmonitored bioaccumulative compounds in bottlenose dolphins from Southern California.

The corresponding article includes a description of
the study goals, samples, and GCxGC/TOF-MS
analytical method. The original PDF version of this library is supplementary
material to the article.
}
\keyword{ package }
